requirements.txt
flask
cassandra-driver
Cluster